Increase length for input endpoint value on registration and output better messages for invalid inputs (#284)
When a microservice instance is registered, an endpoint value (ie network
address) is input.  This value was limited to a length of 64 chars and a
few users have run into this when using an endpoint type of http and using
a URL that was longer than 64 chars.

This code change changes the endpoint values length from 64 to 128.  We
asked users who ran into this issue if 128 would be adequate and they
indicated it would be.

Additonally with this change, we felt we could return a more descriptive
error message when invalid input is specified vs a generic message that
the registration failed.  5 messages are now output when service name is
not specified, service name is too long, endpoint value is too long,
status value is too long, and metadata is too long.