Makefile update for grep on a Mac (#416)
The -x on the grep does not work the same on a Mac so we end up doing all the tasks against the vendor/* sub-directories as well which results in failures. Removing the -x solves the problem on the Mac and I tried against the Dev VM and saw no issues.