Generate sidecar dockerfile for apps for examples testing.
The app Dockerfile.sidecar file has the version number hardcoded.
The examples automated testing builds the images to be pushed to
dockerhub so we want to make sure we are testing with the current
files and not what is uploaded for the previous release. Generate
the Dockerfile.sidecar in 2 ways:
1)  If testing the examples, pull the local release sidecar tar and
extract that.
2)  If not testing the examples, the build.exampleapps target will
generate the dockerfile with the current release version.